[department-of-veterans-affairs/orchid#88] Updates conditional logic and styles veteran information section
