+++ src/applications/ask-a-question/config/pages/contactInformationPage.js
@@ -2,25 +2,15 @@ import set from 'platform/utilities/data/set';
 import { countries } from 'platform/forms/address';
 import fullNameUI from 'platform/forms-system/src/js/definitions/fullName';
 import emailUI from 'platform/forms-system/src/js/definitions/email';
-import currentOrPastDateUI from 'platform/forms-system/src/js/definitions/currentOrPastDate';
 import { confirmationEmailUI } from '../../../caregivers/definitions/caregiverUI';
+import { veteranStatusSection } from './veteranStatusSection';
 
 import fullSchema from '../../0873-schema.json';
 import pageDescription from '../../content/PageDescription';
 
 const countryValues = countries.map(object => object.value);
 const countryNames = countries.map(object => object.label);
-
 const { fullName, email, preferredContactMethod } = fullSchema.definitions;
-
-const {
-  veteranStatus,
-  isDependent,
-  relationshipToVeteran,
-  veteranIsDeceased,
-  dateOfDeath,
-  branchOfService,
-} = fullSchema.properties;
 
 const formFields = {
   preferredContactMethod: 'preferredContactMethod',
@@ -29,71 +19,14 @@ const formFields = {
   email: 'email',
   verifyEmail: 'view:email',
   phoneNumber: 'phoneNumber',
-  veteranStatus: 'veteranStatus',
-  isDependent: 'isDependent',
-  relationshipToVeteran: 'relationshipToVeteran',
-  veteranIsDeceased: 'veteranIsDeceased',
-  dateOfDeath: 'dateOfDeath',
-  branchOfService: 'branchOfService',
   country: 'country',
+  veteranStatusSection: 'veteranStatusSection',
 };
-
-const requireVetRelationship = selectedVeteranStatus =>
-  selectedVeteranStatus === 'behalf of vet' ||
-  selectedVeteranStatus === 'dependent';
-
-const requireServiceInfo = selectedVeteranStatus =>
-  selectedVeteranStatus && selectedVeteranStatus !== 'general';
 
 const contactInformationPage = {
   uiSchema: {
     'ui:description': pageDescription('Your contact info'),
     [formFields.fullName]: fullNameUI,
-    [formFields.veteranStatus]: {
-      'ui:title': 'My message is about benefits/services',
-    },
-    [formFields.isDependent]: {
-      'ui:title': 'Are you the dependent?',
-      'ui:widget': 'yesNo',
-      'ui:required': formData => formData.veteranStatus === 'dependent',
-      'ui:options': {
-        expandUnder: 'veteranStatus',
-        expandUnderCondition: 'dependent',
-      },
-    },
-    [formFields.relationshipToVeteran]: {
-      'ui:title': 'Your relationship to the Veteran',
-      'ui:required': formData => requireVetRelationship(formData.veteranStatus),
-      'ui:options': {
-        expandUnder: 'veteranStatus',
-        expandUnderCondition: requireVetRelationship,
-      },
-    },
-    [formFields.veteranIsDeceased]: {
-      'ui:title': 'Is the Veteran deceased?',
-      'ui:widget': 'yesNo',
-      'ui:required': formData => requireVetRelationship(formData.veteranStatus),
-      'ui:options': {
-        expandUnder: 'veteranStatus',
-        expandUnderCondition: requireVetRelationship,
-      },
-    },
-    [formFields.dateOfDeath]: {
-      ...currentOrPastDateUI('Date of Death if known'),
-      ...{
-        'ui:options': {
-          expandUnder: 'veteranStatus',
-          hideIf: formData => !formData.veteranIsDeceased,
-        },
-      },
-    },
-    [formFields.branchOfService]: {
-      'ui:title': 'Branch of service',
-      'ui:required': formData => requireServiceInfo(formData.veteranStatus),
-      'ui:options': {
-        hideIf: formData => !requireServiceInfo(formData.veteranStatus),
-      },
-    },
     [formFields.email]: set(
       'ui:required',
       (formData, _index) => formData.preferredContactMethod === 'email',
@@ -107,23 +40,17 @@ const contactInformationPage = {
       'ui:title': 'How should we get in touch with you?',
       'ui:widget': 'radio',
     },
+    [formFields.veteranStatusSection]: { ...veteranStatusSection.uiSchema },
   },
   schema: {
     type: 'object',
     required: [
       formFields.preferredContactMethod,
       formFields.fullName,
-      formFields.veteranStatus,
       formFields.country,
     ],
     properties: {
       [formFields.fullName]: fullName,
-      [formFields.veteranStatus]: veteranStatus,
-      [formFields.isDependent]: isDependent,
-      [formFields.relationshipToVeteran]: relationshipToVeteran,
-      [formFields.veteranIsDeceased]: veteranIsDeceased,
-      [formFields.dateOfDeath]: dateOfDeath,
-      [formFields.branchOfService]: branchOfService,
       [formFields.email]: email,
       [formFields.verifyEmail]: {
         type: 'string',
@@ -135,6 +62,7 @@ const contactInformationPage = {
         enumNames: countryNames,
       },
       [formFields.preferredContactMethod]: preferredContactMethod,
+      [formFields.veteranStatusSection]: { ...veteranStatusSection.schema },
     },
   },
 };
