+++ src/applications/ask-a-question/content/SectionHeader.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function SectionHeader(headerText) {
+  return (
+    <div>
+      <h3>{headerText}</h3>
+    </div>
+  );
+}
