+++ src/applications/ask-a-question/tests/data/maximal-test.json
@@ -8,8 +8,14 @@
       "last": "Doe",
       "suffix": "IV"
     },
-    "relationshipToVeteran": "On behalf of a Veteran",
-    "branchOfService": "Army",
+    "veteranStatusSection": {
+      "veteranStatus": "For the dependent of a Veteran",
+      "isDependent": true,
+      "relationshipToVeteran": "Daughter",
+      "veteranIsDeceased": true,
+      "dateOfDeath": "2000-06-04",
+      "branchOfService": "Army"
+    },
     "preferredContactMethod": "email",
     "email": "jane.doe@va.gov",
     "view:email": "jane.doe@va.gov",
