+++ src/applications/ask-a-question/tests/data/minimal-test.json
@@ -4,7 +4,9 @@
       "first": "Jane",
       "last": "Doe"
     },
-    "relationshipToVeteran": "A general question",
+    "veteranStatusSection": {
+      "veteranStatus": "A general question"
+    },
     "preferredContactMethod": "email",
     "email": "jane.doe@va.gov",
     "view:email": "jane.doe@va.gov",
